% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrega.R
\name{aggrega}
\alias{aggrega}
\title{aggrega}
\usage{
aggrega(x, aggFun, max.na, rle.check, max.size.block.na, seasonalCheck)
}
\arguments{
\item{x}{Un oggetto ClimateData.}

\item{aggFun}{Funzione per aggregare i dati in x: media o somma.}

\item{max.na}{Massimo numero di NA ammessi in x (integer).}

\item{rle.check}{Verificare o meno la presenza di blocchi di NA contigui? (Logical).}

\item{max.size.block.na}{Massimo numero di NA contigui (Integer). Questo parametro deve essere utilizzato con rle.check=TRUE.}

\item{seasonalCheck}{Verificare che tutte le stagioni siano valide? (Logical).}
}
\value{
double
}
\description{
La funzione aggrega viene utilizzata sia da apply.monthly che da apply.yearly per
aggregare (sommare o mediare, in base al parametro) i dati di precipitazione/temperatura.
Nel caso dell'aggregazione annuale è necessario passare seasonal=TRUE in modo che venga verificata la validità di tutte le
stagioni in un anno. Una stagione si considera valida quando sono validi almeno due mesi su tre.
}
