% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrega.R
\name{aggregaCD}
\alias{aggregaCD}
\alias{aggregaCD.default}
\alias{aggregaCD.daily}
\alias{aggregaCD.monthly}
\alias{aggregaCD.yearly}
\title{aggregaCD}
\usage{
aggregaCD(x, max.na, rle.check, max.size.block.na, ...)

\method{aggregaCD}{default}(x)

\method{aggregaCD}{daily}(x, max.na, rle.check, max.size.block.na)

\method{aggregaCD}{monthly}(x, max.na, rle.check, max.size.block.na, ignore.par, seasonal)

\method{aggregaCD}{yearly}(x, max.na, rle.check, max.size.block.na, ignore.par, seasonal)
}
\arguments{
\item{x}{Un oggetto ClimateData.}

\item{max.na}{Massimo numero di NA che sono disposto a tollerare (integer)}

\item{rle.check}{Se TRUE viene verificato che all'interno di x non vi siano blocchi di NA di lunghezza superiore a "max.size.block.na"?}

\item{max.size.block.na}{Massimo numero di NA contigui (integer). Da usare con rle.check=TRUE.}

\item{...}{Uno o piu' parametri in base alla tipologia di x.}

\item{ignore.par}{Se TRUE viene ignorato il tipo di parametro (precipitazione o temperatura) in fase di aggregazione.}

\item{seasonal}{Se TRUE viene controllata la validita' delle stagioni. Una stagione e' valida se contiene al massimo un mese NA.}
}
\value{
Un oggetto ClimateData.
}
\description{
Funzioni per aggregare le serie di dati climatici
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Funzione di default

\item \code{daily}: Aggrega un ClimateData con serie giornaliere

\item \code{monthly}: Aggrega un ClimateData con serie mensili

\item \code{yearly}: Aggrega un ClimateData con serie annuali.

aggregaCD.yearly prende dati annuali e restituisce ancora un oggetto con classe "yearly".
Questa funzione viene utilizzata per creare un valore singolo (somma o media) di valori annuali
Per la precipitazione sono possibili due operazioni:
\itemize{
\item somma (se calcolo l'aggregato annuale)
\item media: l'operazione media viene utilizzata per costruire il climatologico mensile (ad esempio il climatologico del mese di gennaio)
}
}}

