% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateData.R
\name{all_NA}
\alias{all_NA}
\title{all_NA}
\usage{
all_NA(x)
}
\arguments{
\item{x}{Un oggetto ClimateData}
}
\value{
Logical
}
\description{
Verifica se un oggetto ClimateData sia composto di soli NA
Utile ad esempio, nel calcolo dei climatologici mensili.
}
