% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalie.R
\name{anomalie}
\alias{anomalie}
\alias{anomalie.monthly}
\alias{anomalie.yearly}
\title{anomalie}
\usage{
anomalie(x, climatol)

\method{anomalie}{monthly}(x, climatol)

\method{anomalie}{yearly}(x, climatol)
}
\arguments{
\item{x}{Oggetto ClimateData per serie mensili o annuali.}

\item{climatol}{Valori climatologici rispetto a cui calcolare l'anomalia.}
}
\value{
Oggetto ClimateData con le anomalie mensili o annuali di temperatura o precipitazione.
}
\description{
Funzioni per il calcolo delle anomalie mensili e annuali.
}
\section{Methods (by class)}{
\itemize{
\item \code{monthly}: Calcolo anomalie mensili.

\item \code{yearly}: Calcolo delle anomalie annuali.
}}

