% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateData.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.ClimateData}
\title{as.data.frame.ClimateData}
\usage{
as.data.frame(x, ...)

\method{as.data.frame}{ClimateData}(x)
}
\arguments{
\item{x}{Un oggetto ClimateData.}

\item{...}{Uno o piu' parametri per la funzione as.data.frame.}
}
\value{
Un dataframe.
}
\description{
Trasforma un oggetto ClimateData in un data.frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{ClimateData}: Crea un dataframe da un oggetto ClimateData.
}}

