% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateData.R
\name{checkParam}
\alias{checkParam}
\title{checkParam}
\usage{
checkParam(x)
}
\arguments{
\item{x}{Stringa di caratteri}
}
\value{
Una stringa di caratteri
}
\description{
Funzione per verificare il parametro passato a ClimateData
}
