% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_series_validity.R
\name{checkSeriesValidity2}
\alias{checkSeriesValidity2}
\title{checkSeriesValidity2}
\usage{
checkSeriesValidity2(
  x,
  percentuale.anni.validi,
  max.size.block.na,
  lunghezza.minima.serie = 10
)
}
\arguments{
\item{x}{Un oggetto ClimateData con serie annuali}

\item{percentuale.anni.validi}{Percentuale di anni validi nella serie (numeric)}

\item{max.size.block.na}{Massimo numero di NA consecutivi ammesso (integer)}

\item{lunghezza.minima.serie}{Numero di anni che devono essere validi (non NA) (integer)}
}
\value{
Un oggetto ClimateData con le sole serie valide o NULL
}
\description{
NON IMPLEMENTATA, probabilmente inutile
Verifica la validità di una serie annuale secondo i controlli applicati
per la selezione delle serie SCIA utilizzate per la stima dei trend.
}
\details{
\itemize{
\item L'86% dei dati deve essere valido
\item Un blocco di NA può al massimo essere lungo 4 anni
\item Gli ultimi 3 anni devono essere validi, non NA
}
}
