% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_series_validity.R
\name{checkSeriesValidity_default_values}
\alias{checkSeriesValidity_default_values}
\title{checkSeriesValidity_default_values}
\usage{
checkSeriesValidity_default_values(...)
}
\arguments{
\item{...}{Non implementato}
}
\description{
checkSeriesValidity_default_values
}
