% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateData.R
\name{ClimateData}
\alias{ClimateData}
\alias{ClimateData.data.frame}
\alias{ClimateData.xts}
\title{ClimateData}
\usage{
ClimateData(x, param)

\method{ClimateData}{data.frame}(x, param = NULL)

\method{ClimateData}{xts}(x, param = NULL)
}
\arguments{
\item{x}{Un dataframe o un oggetto xts da passare a ClimateData.}

\item{param}{Una stringa di testo (ammesse minuscole e maiuscole) per specificare se si tratta di dati di temperatura o precipitazione.
Per definire serie di dati di temperatura sono ammesse le seguenti stringhe: "tas","tasmax","tasmin", "tmax","tmin","tmean".
Per definire serie di dati di precipitazione sono ammesse le seguenti stringhe: "pr","prpc","prcp", "prec".}
}
\value{
Un oggetto di tipo ClimateData.
}
\description{
Funzioni per creare oggetti Climate Data
}
\details{
Un oggetto ClimateData e' un oggetto xts con un attributo parametro.

Un oggetto ClimateData puo' essere costruito a partire da un data.frame che deve contenere obbligatoriamente la colonna che identifica gli anni: yy o year.
Altre colonne opzionali sono: mm (o month) se si tratta di serie di dati mensili; dd (o day) se si tratta di serie di dati giornalieri.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Crea un oggetto ClimateData a partire da un data.frame.

\item \code{xts}: Crea un oggetto ClimateData partendo da un oggetto xts.
}}

