% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatologici.R
\name{climatologiciMensili}
\alias{climatologiciMensili}
\title{climatologiciMensili}
\usage{
climatologiciMensili(x, yearS, yearE, max.na, rle.check, max.size.block.na)
}
\arguments{
\item{x}{Oggetto ClimateData con serie mensili}

\item{yearS}{Anno di inizio per il calcolo del climatologico (start year)}

\item{yearE}{Anno di fine per il calcolo del climatologico (end year)}

\item{max.na}{Massimo numero di NA ammessi nella serie.}

\item{rle.check}{Controllare la presenza di blocchi di NA? (logical)}

\item{max.size.block.na}{Numero massimo di NA consecutivi (da usare con rle.check=TRUE)}
}
\value{
Oggetto ClimateData con valori climatologici mensili.
}
\description{
Utilizza la funzione "calcolaClimatologico" per il calcolo dei valori climatologici mensili e annuali secondo i criteri del WMO (vedi seguito).
 Calcolo dei valori climatologici secondo i criteri del WMO.
 \itemize{
 \item Dai dati giornalieri (oggetto ClimateData daily) calcolare l'aggregato mensile con aggregateCD
 \item Utilizzare la funzione climatologiciMensili per ottenere il climatologico di ogni mese. I dati cosi' ottenuti sono oggetti ClimateData monthly.
 \item Utilizzare aggregaeCD (per l'esattezza: aggregaCD.monthly) con max.na=0 per ottenere il climatologico annuale.
 }
}
\details{
Attenzione: l'output e' un tibble e non un ClimateData non per errore. Se fosse un ClimateData avrebbe periodicita' mensile
 che pero' creerebbe problemi ad aggregaCD per calcolare il climatologico annuale. Il nodo corretto di procedere e' quello di
 prendere il tibble di output, eliminare la colonna dei mesi e creare un oggetto ClimateData. Questo avra' periodicita' annuale (mancando la colonna dei mesi).
 L'oggetto ClimateData puo' essere poi passato ad aggregateCD per creare il climatologico annuale.
}
