% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatologici.R
\name{climatologicoAnnuale}
\alias{climatologicoAnnuale}
\title{climatologicoAnnuale}
\usage{
climatologicoAnnuale(x)
}
\arguments{
\item{x}{Un tibble con colonne yy e mm. Si tratta del tibble restituito da climatologiciMensili.}
}
\value{
Un data.frame con i valori climatologici annuali.
}
\description{
Funzione per il calcolo del climatologico annuale
}
