% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateData.R
\name{is.ClimateData}
\alias{is.ClimateData}
\title{is.ClimateData}
\usage{
is.ClimateData(x)
}
\arguments{
\item{x}{Un oggetto R.}
}
\value{
Logical
}
\description{
Restituisce TRUE se si tratta di un oggetto ClimateData, altrimenti FALSE.
}
