% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalie.R
\name{OpWithClimatol}
\alias{OpWithClimatol}
\title{OptWithClimatol}
\usage{
OpWithClimatol(fun)
}
\arguments{
\item{fun}{Funzione da applicare.}
}
\description{
Differenza tra ciascuna serie mensile e i valori climatologici mensili
}
