% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalie.R
\name{ricostruisciSerie}
\alias{ricostruisciSerie}
\alias{ricostruisciSerie.monthly}
\alias{ricostruisciSerie.yearly}
\title{ricostruisciSerie}
\usage{
ricostruisciSerie(x, climatol)

\method{ricostruisciSerie}{monthly}(x, climatol)

\method{ricostruisciSerie}{yearly}(x, climatol)
}
\arguments{
\item{x}{Serie di anomalie mensili o annuali.}

\item{climatol}{Valori climatologici mensili o annuali.}
}
\value{
ClimateData con le serie mensili o annuali di temperatura o precipitazione.
}
\description{
Operazione inversa a quella effettuata da anomalie
Prende un oggetto prodotto da anomalie e ricostruisce le serie.
}
\section{Methods (by class)}{
\itemize{
\item \code{monthly}: Ricostruisce le serie mensili.

\item \code{yearly}: Ricostruisce le serie annuali.
}}

