% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrega.R, R/verifica_blocchi_NA.R
\name{verifica_blocchi_NA}
\alias{verifica_blocchi_NA}
\title{verifica_blocchi_NA:}
\usage{
verifica_blocchi_NA(x, max.size.block.na)

verifica_blocchi_NA(x, max.size.block.na)
}
\arguments{
\item{x}{Un vettore numerico.}

\item{max.size.block.na}{Lunghezza massima accettabile di NA contigui in x (integer).}
}
\value{
list

list
}
\description{
Funzione per identificare blocchi di NA contigui.

Funzione per identificare blocchi di NA contigui.
}
\details{
verifica_blocchi_NA restituisce una lista di due elementi:
\itemize{
\item "ris": boolean pari a TRUE se il vettore x contiene almeno un blocco di NA di lunghezza superiore a max.size.block.na.
\item "nb": number of blocks, riporta il numero di blocchi in x che superano max.size.block.na.
}

verifica_blocchi_NA restituisce una lista di due elementi:
\itemize{
\item "ris": boolean pari a TRUE se il vettore x contiene almeno un blocco di NA di lunghezza superiore a max.size.block.na.
\item "nb": number of blocks, riporta il numero di blocchi in x che superano max.size.block.na.
}
}
